// libs/logic-apps-shared/src/utils/src/lib/models/logicApps.ts
export type ActionDefinition =
  | ApiConnectionAction
  | ApiConnectionWebhookAction
  | ChildWorkflow
  | ComposeAction
  | FunctionAction
  | HttpAction
  | HttpWebhookAction
  | WaitAction
  | Integration.FlatFileDecodingAction
  | Integration.FlatFileEncodingAction
  | Integration.IntegrationAccountArtifactLookupAction
  | Integration.XmlValidationAction
  | Integration.XsltAction;

export type OperationDefinition = ActionDefinition | TriggerDefinition;

export interface Actions {
  [action: string]: ActionDefinition;
}

/* API Connection actions and triggers */
export interface ApiConnectionAction extends Action {
  inputs: ApiConnectionInputs;
}

export interface ApiConnectionTrigger extends RecurrenceTrigger {
  inputs: ApiConnectionInputs;
}
// ...
export interface ApiConnectionWebhookAction extends Action {
  inputs: ApiConnectionWebhookInputs;
}

export interface ApiConnectionWebhookTrigger extends Trigger {
  inputs: ApiConnectionWebhookInputs;
}
// ...
export interface ChildWorkflow extends Action {
  inputs: ChildWorkflowInputs;
}
// ...
export interface ComposeAction extends Action {
  inputs: any;
}

/* Condition types */

export interface Condition {
  dependsOn?: string;
  expression?: string;
}

/* HTTP action and trigger types */

export interface HttpAction extends Action {
  inputs: HttpInputs;
  metadata?: Record<string, string>;
}

export interface HttpTrigger extends RecurrenceTrigger {
  inputs: HttpInputs;
  metadata?: Record<string, string>;
}
// ...
export interface HttpWebhookAction extends Action {
  inputs: HttpWebhookInputs;
}

export interface HttpWebhookTrigger extends Trigger {
  inputs: HttpWebhookInputs;
}

export interface HttpWebhookInputs extends ManualActionInputs {
  subscribe: HttpInputs;
  unsubscribe: HttpInputsOptional;
}

/* Manual action input types */

export interface ManualActionInputs {
  schema?: any;
}

/* Manual trigger types */

export interface ManualTrigger extends Trigger {
  inputs?: ManualTriggerInputs;
}

export interface ManualTriggerInputs {
  method?: string;
  relativePath?: string;
  schema?: any;
}

/* Function types */

export interface FunctionAction extends Action {
  inputs: FunctionInputs;
}
// ...
export interface RecurrenceTrigger extends Trigger {
  recurrence?: Recurrence;
}
// ...
export type TriggerDefinition =
  | ApiConnectionTrigger
  | ApiConnectionWebhookTrigger
  | HttpTrigger
  | HttpWebhookTrigger
  | ManualTrigger
  | RecurrenceTrigger;

export interface Triggers {
  [trigger: string]: TriggerDefinition;
}

/* Wait action types */

export interface WaitAction extends Action {
  inputs: WaitInputs;
}